# BookingWebApp/src/app/modules/data-access/bookingwebapi.definition.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BookingWebAPI",
    "version": "1.0"
  },
  "servers" : [ {
    "url": "https://api.ezbooking.com:8000",
    "description": "BookingWebAPI dev"
  }],
  "paths": {
...
    "/api/Users/ConfirmUser": {
      "get": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookingWebAPIUserViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingWebAPIUserViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingWebAPIUserViewModel"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRegistrationViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRegistrationViewModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRegistrationViewModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookingWebAPIUserViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingWebAPIUserViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingWebAPIUserViewModel"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookingWebAPIUserViewModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ConfirmRegistrationViewModel": {
        "required": [
          "password"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "token": {
            "type": "string",
            "format": "uuid"
          },
          "password": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}