# BookingWebApp/src/app/modules/data-access/model/bookingWebAPIUserViewModel.ts
/**
 * BookingWebAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BookingWebAPIUserViewModel { 
    id?: string;
    userName?: string | null;
    email?: string | null;
    firstName?: string | null;
    lastName?: string | null;
}